+++ source/github/events/create_installation.ts
@@ -2,29 +2,18 @@ import * as express from "express"
 
 import { Installation } from "../events/types/integration_installation_created.types"
 
-import { GitHubInstallation } from "../../db"
 import { getDB } from "../../db/getDB"
+import generateInstallation from "../../testing/installationFactory"
 
 export async function createInstallation(installationJSON: Installation, _: express.Request, res: express.Response) {
-  const installation: GitHubInstallation = {
-    id: installationJSON.id,
-    repos: {},
-    rules: {
-      pull_request: "dangerfile.js",
-    },
-    scheduler: {},
-    settings: {
-      env_vars: [],
-      ignored_repos: [],
-      modules: [],
-    },
-    tasks: {},
-  }
-
-  // Default to no runnerRules
-
-  res.status(200).send("Creating new installation.")
+  const installation = generateInstallation({ iID: installationJSON.id })
 
   const db = getDB()
-  await db.saveInstallation(installation)
+  const existingInstallation = await db.getInstallation(installation.iID)
+  if (existingInstallation) {
+    res.status(204).send("Did not create new installation, it already existed.")
+  } else {
+    await db.saveInstallation(installation)
+    res.status(200).send("Creating new installation.")
+  }
 }