+++ source/peril.ts
@@ -7,11 +7,12 @@ import {
   HYPER_ACCESS_KEY,
   MONGODB_URI,
   PERIL_WEBHOOK_SECRET,
+  PUBLIC_API_ROOT_URL,
   PUBLIC_FACING_API,
   validateENVForPerilServer,
 } from "./globals"
 
-import prDSLRunner from "./api/pr/dsl"
+import setupPublicAPI from "./api/api"
 import logger from "./logger"
 import { hyperUpdater } from "./routing/hyper_updater"
 import webhook from "./routing/router"
@@ -29,7 +30,6 @@ const peril = () => {
 
   const app = express()
   app.set("port", process.env.PORT || 5000)
-  app.set("view engine", "ejs")
   app.use(xhub({ algorith: "sha1", secret: PERIL_WEBHOOK_SECRET }))
   app.use(bodyParser.json())
   app.use(express.static("public"))
@@ -41,20 +41,22 @@ const peril = () => {
     startTaskScheduler()
   }
 
-  if (PUBLIC_FACING_API) {
-    logger.info("This is a public facing Peril instance.")
-    app.get("/api/v1/pr/dsl", prDSLRunner)
-  }
-
   if (HYPER_ACCESS_KEY) {
     // You need to set up a dockerhub webhook that is your peril address
     // with you hyper access key as a secret for your URL.
     //
     // This is mainly so you don't need to remember to do a `hyper pull` after
-    // an update.
+    // an update to Peril.
     const url = `/api/v1/webook/dockerhub/${HYPER_ACCESS_KEY}`
     logger.info(`Started up the dockerhub webhook on ${url}`)
     app.post(url, hyperUpdater)
+  }
+
+  // This should go last
+  if (PUBLIC_FACING_API) {
+    logger.info("This is a public facing Peril instance.")
+    logger.info(`- API Root: ${PUBLIC_API_ROOT_URL}`)
+    setupPublicAPI(app)
   }
 
   // Start server