+++ source/tasks/runTask.ts
@@ -1,5 +1,5 @@
 import { getTemporaryAccessTokenForInstallation } from "../api/github"
-import { dangerRepresentationforPath, dsl } from "../danger/danger_run"
+import { dangerRepresentationForPath, dsl } from "../danger/danger_run"
 import { runDangerForInstallation } from "../danger/danger_runner"
 import { DangerfileReferenceString, GitHubInstallation } from "../db/index"
 import { getGitHubFileContents } from "../github/lib/github_helpers"
@@ -13,7 +13,7 @@ const error = (message: string) => {
 
 export const runTask = async (installation: GitHubInstallation, rules: DangerfileReferenceString, data: any) => {
   // This is duped with runJob
-  const rep = dangerRepresentationforPath(rules)
+  const rep = dangerRepresentationForPath(rules)
   if (rep.repoSlug === undefined) {
     if (DATABASE_JSON_FILE) {
       // If you don't provide a repo slug, assume that the
@@ -30,7 +30,7 @@ export const runTask = async (installation: GitHubInstallation, rules: Dangerfil
       data,
     },
   }
-  const token = await getTemporaryAccessTokenForInstallation(installation.id)
+  const token = await getTemporaryAccessTokenForInstallation(installation.iID)
   const dangerfile = await getGitHubFileContents(token, rep.repoSlug!, rep.dangerfilePath, rep.branch)
   return runDangerForInstallation(dangerfile, rep.dangerfilePath, null, dsl.import, installation, dangerDSL)
 }